fix: ignore messages when permissions are valid

+++ main.py
@@ -3,6 +3,7 @@ import os
 import re
 from pattern import pattern
 
+
 class MyClient(discord.Client):
     async def on_ready(self):
         print(f'Logged on as {self.user}')
@@ -11,12 +12,16 @@ class MyClient(discord.Client):
         if (message.author.bot):
             return
         if message.channel.id == int(os.environ.get("CHANNEL_ID")):
+            if (message.author.guild_permissions.manage_messages or message.author.guild_permissions.administrator):
+                return
             print("Found message in right channel")
             if not re.search(pattern, message.content):
-              await message.reply("So sorry, but only `twitch.tv` and `kick.com` links are allowed in this channel.")
-              await message.delete()
-              webhook = discord.Webhook.from_url(os.environ.get("WH_URL"), client=self)
-              await webhook.send(content=f"Deleted message for failing link test:\n\n> {message.content}")
+                await message.reply("So sorry, but only `twitch.tv` and `kick.com` links are allowed in this channel.")
+                await message.delete()
+                webhook = discord.Webhook.from_url(
+                    os.environ.get("WH_URL"), client=self)
+                await webhook.send(content=f"Deleted message for failing link test:\n\n> {message.content}")
+
 
 intents = discord.Intents.default()
 intents.message_content = True
